enhancement: enhancement: add GeoCoordinate Default
