+++ src/Core/Timers/IObservableTimer.cs
@@ -1,8 +1,6 @@
-using System;
-
 namespace Rocket.Surgery.Airframe;
 
 /// <summary>
 /// Represents an <see cref="ITimer"/> that is observable.
 /// </summary>
-public interface IObservableTimer : ITimer, IObservable<TimeSpan>;
+public interface IObservableTimer : ITimer;
