+++ test/Composition.Tests/Composition.Tests.csproj
@@ -6,21 +6,6 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <PackageReference Include="coverlet.collector" />
-        <PackageReference Include="coverlet.msbuild" />
-        <PackageReference Include="FluentAssertions" />
-        <PackageReference Include="Microsoft.NET.Test.Sdk" />
-        <PackageReference Include="ReactiveUI" />
-        <PackageReference Include="Rocket.Surgery.Extensions.Testing.Fixtures" />
-        <PackageReference Include="Splat" />
-        <PackageReference Include="Splat.DryIoc" />
-        <PackageReference Include="System.ComponentModel" />
-        <PackageReference Include="System.Reactive" />
-        <PackageReference Include="xunit" />
-        <PackageReference Include="xunit.runner.visualstudio" />
-    </ItemGroup>
-
-    <ItemGroup>
       <ProjectReference Include="..\..\src\Composition\Composition.csproj" />
     </ItemGroup>
